+++ Runtime/Components/AutoDestroy.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using UnityEngine.Events;
 
 namespace NekoLib.Components
 {
@@ -6,17 +7,28 @@ namespace NekoLib.Components
     [AddComponentMenu("NekoLib/Auto Destroy")]
     public class AutoDestroy : MonoBehaviour
     {
-        [SerializeField, Min(0f), Tooltip("Time in seconds before the object is destroyed")] private float _lifetime = 5f;
+        [SerializeField, Min(0f), Tooltip("Time in seconds before the object is destroyed")]
+        private float _destroyAfter = 5f;
+
+        [SerializeField, Space(6), Tooltip("Event triggered just before destruction")]
+        private UnityEvent _onBeforeDestroy;
+        public UnityEvent OnBeforeDestroy => _onBeforeDestroy;
 
         private void Start()
         {
-            if (_lifetime <= 0f)
+            if (_destroyAfter <= 0f)
             {
-                Debug.LogWarning("[AutoDestroy] Lifetime must be greater than zero. Object will not be destroyed.");
+                Debug.LogWarning("[AutoDestroy] Destroy delay must be greater than zero. Object will not be destroyed.", this);
+                enabled = false;
                 return;
             }
 
-            Destroy(gameObject, _lifetime);
+            Destroy(gameObject, _destroyAfter);
+        }
+
+        private void OnDestroy()
+        {
+            _onBeforeDestroy?.Invoke();
         }
     }
 }