+++ Runtime/Components/Editor/AutoDestroyEditor.cs
@@ -1,0 +1,19 @@
+#if UNITY_EDITOR
+using UnityEditor;
+
+namespace NekoLib.Components
+{
+    [CustomEditor(typeof(AutoDestroy))]
+    public class AutoDestroyEditor : Editor
+    {
+        public override void OnInspectorGUI()
+        {
+            serializedObject.Update();
+
+            DrawPropertiesExcluding(serializedObject, "m_Script");
+
+            serializedObject.ApplyModifiedProperties();
+        }
+    }
+}
+#endif