+++ Runtime/Utilities/EventUtils.cs
@@ -1,0 +1,12 @@
+using System;
+using UnityEngine.Events;
+
+namespace NekoLib.Utilities
+{
+    public static partial class Utils
+    {
+
+    }
+
+    [Serializable] public class FloatEvent : UnityEvent<float> { }
+}